Moved attachment src, filename, and mapping to main() function
The previous version of the script built a list of all attachment "src" elements, attachment filenames, and then a map of the two every single time the write_mms_messages function was called. This version moves that attachment processing and mapping to the main() function so it is only run once for the entire script. This required me to pass the src_filename_map list to the write_sms_messages and write_mms_messages functions, but overall results in a drastic reduction in processing time for the script. For my data set, the previous version took around 15-16 minutes to process 5312 messages, 145 images, and 23 contact cards. This commit reduces that processing time down to around 12 seconds.